pyproject.toml
[project]
name = "asl-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "mediapipe==0.10.14",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "tqdm>=4.67.3",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]
